Upload Truncate Class Implementation and Header
The truncate class is designed to assist with fixed length management by truncating an input string, from the right, not to exceed a size of maximum length.  If a string must be modified, the modifyString() function will edit the input string.  If a field or parameter can't exceed a size but the string must not be changed, then the truncatedString() function will provide a truncated version of the string with out editing the input string.